fixed parsing error
the experiment failed to run in the browser, because of the format of the runif statement in the trial sequences.

These had the format: runif [D] != "w"

These quote marks got automatically excepted out with a backslash, which was interpeted in the browser as a javascript end of line character.